+++ app/gas/admin.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 
 from django.contrib import admin
 
-from .models import GasAdvisory, GasGuzzler, GasHistory, GasProfile
+from .models import GasAdvisory, GasGuzzler, GasProfile
 
 
 class GeneralAdmin(admin.ModelAdmin):
@@ -11,7 +11,6 @@ class GeneralAdmin(admin.ModelAdmin):
     ordering = ['-id']
 
 
-admin.site.register(GasHistory, GeneralAdmin)
 admin.site.register(GasAdvisory, GeneralAdmin)
 admin.site.register(GasGuzzler, GeneralAdmin)
 admin.site.register(GasProfile, GeneralAdmin)
