+++ app/perftools/admin.py
@@ -1,0 +1,10 @@
+from django.contrib import admin
+from .models import JSONStore
+
+
+class GeneralAdmin(admin.ModelAdmin):
+
+    ordering = ['-id']
+
+
+admin.site.register(JSONStore, GeneralAdmin)
